pyproject.toml
[project]
name = "twiga-wikipedia-chunker"
version = "0.1.0"
description = "Convert a syllabus into chunks from Wikipedia articles"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain-openai>=0.3.0",
    "langchain-together>=0.3.0",
    "langchain-unstructured>=0.1.6",
    "langchain>=0.3.14",
    "matplotlib>=3.10.0",
    "pillow>=11.1.0",
    "pymupdf>=1.25.2",
    "python-dotenv>=1.0.1",
    "together>=0.2.4",
    "typer>=0.15.1",
    "wikipedia>=1.4.0",
]
